# backend/swagger.yaml
openapi: 3.0.3
info:
  title: ERP System Backend API
  description: |
    Complete API documentation for the ERP System Backend microservices.
    
    ## Services
    - **Auth Service**: Authentication and user management
    - **Organization Service**: Organizations, companies, and locations
    - **Inventory Service**: Products and stock management
    - **License Service**: Application licenses and assignments
    - **Subscription Service**: Subscription plans and billing
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: ERP System Team
    email: support@erpsystem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Auth Service (Development)
  - url: http://localhost:8082/api/v1
    description: Organization Service (Development)
  - url: http://localhost:8083/api/v1
    description: Inventory Service (Development)
  - url: http://localhost:8084/api/v1
    description: License Service (Development)
  - url: http://localhost:8085/api/v1
    description: Subscription Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 400
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    PaginationResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
    BaseModel:
      type: object
      properties:
        id:
          type: string
          format: objectid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: objectid
        updated_by:
          type: string
          format: objectid
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          format: objectid
          nullable: true
...
    Organization:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            name:
              type: string
            legal_name:
              type: string
            domain:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
            website:
              type: string
            tax_id:
              type: string
            registration_number:
              type: string
            industry:
              type: string
            company_size:
              type: string
              enum: [small, medium, large, enterprise]
            status:
              type: string
              enum: [active, inactive, suspended, trial, cancelled]
            is_active:
              type: boolean
            billing_email:
              type: string
              format: email
            max_users:
              type: integer
            max_companies:
              type: integer
            max_locations:
              type: integer
            current_users:
              type: integer
            current_companies:
              type: integer
            current_locations:
              type: integer
            storage_used_gb:
              type: number
              format: float
            storage_limit_gb:
              type: number
              format: float
    CreateOrganizationRequest:
      type: object
      required:
        - name
        - legal_name
        - domain
        - email
        - billing_email
      properties:
        name:
          type: string
        legal_name:
          type: string
        domain:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
        tax_id:
          type: string
        registration_number:
          type: string
        industry:
          type: string
        company_size:
          type: string
          enum: [small, medium, large, enterprise]
        billing_email:
          type: string
          format: email
        max_users:
          type: integer
        max_companies:
          type: integer
        max_locations:
          type: integer
        storage_limit_gb:
          type: number
          format: float
...
    License:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            organization_id:
              type: string
              format: objectid
            application_id:
              type: string
              format: objectid
            license_key:
              type: string
            license_type:
              type: string
              enum: [perpetual, subscription, trial]
            status:
              type: string
              enum: [active, expired, suspended, revoked]
            max_users:
              type: integer
            max_devices:
              type: integer
            current_users:
              type: integer
            current_devices:
              type: integer
            valid_from:
              type: string
              format: date-time
            valid_until:
              type: string
              format: date-time
    CreateLicenseRequest:
      type: object
      required:
        - application_id
        - license_type
      properties:
        application_id:
          type: string
        license_type:
          type: string
          enum: [perpetual, subscription, trial]
        max_users:
          type: integer
        max_devices:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
    
    # Subscription Schemas
...
paths:
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: Get paginated list of organizations with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, trial, cancelled]
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    
    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
        '400':
          description: Validation error or creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /organizations/{org_id}/licenses:
    get:
      tags:
        - Licenses
      summary: List licenses by organization
      description: Get paginated list of licenses for an organization
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Licenses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    
    post:
      tags:
        - Licenses
      summary: Create license
      description: Create a new license for an organization
      security:
        - BearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLicenseRequest'
      responses:
        '201':
          description: License created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/License'
  /licenses/{id}:
    get:
      tags:
        - Licenses
      summary: Get license
      description: Retrieve license by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: License retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/License'
  /licenses/{id}/suspend:
    post:
      tags:
        - Licenses
      summary: Suspend license
      description: Suspend an active license
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: License suspended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /applications:
    get:
      tags:
        - Applications
      summary: List applications
      description: Get list of available applications (public endpoint)
      responses:
        '200':
          description: Applications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    
    post:
      tags:
        - Applications
      summary: Create application
      description: Create a new application (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - code
              properties:
                name:
                  type: string
                code:
                  type: string
                description:
                  type: string
                version:
                  type: string
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  
  # ===============================
  # SUBSCRIPTION SERVICE ENDPOINTS
  # ===============================
...